requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==2.3.3
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
outcome==1.3.0
packaging==23.2
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==11.0.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
schedule==1.2.1
selenium==4.14.0
slack-bolt==1.18.0
slack-sdk==3.23.0
slackclient==2.9.4
slackeventsapi==3.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
